feat(order_history): update order history UI and add pagination
This commit adds a new template: `order_history.html`, which combines the old `kitchen_history` and `bar_history` templates into one unifed view. It introduced category filtering and pagination to make browsing the database faster. The report button was removing because this feature has been moved to a different section of the project.