Update logic discount (#50)
* fixbug: fix bug with price_shot

Fix bug in save method. When user first time create model the field price_snapshot is string and it cause bug because we cannot comare string with int

* feat: add discount function and update related models

Add a new method  that update the  field in the Bill model.
This method recalculates price for all related models: OrerItem and OrderItemAddition.It should be called whenever user applies or modifies a discount in the Bill to ensure price consistency across all dependet records.

* test: add basic test for add_discount method

* fix: handle None value in add_discount totals

This bug caused the method fail when either  and  had a value of None

* feat: add management commands for data maintenance and cleanup

Added two new django commands to assist with databse maintenance and data integrity:
*  removes invalid or empty orders and their associated bills to clean up fragmented data
*  recalculates the discount distribution across all existing bills, ensuring line-item values (line_subtotal, line_discount_amount, line_final_total) are consistent and accurate

* feat: add atomic transaction to create_order view

Ensures atomicity in the order creation process.This functionality prevent user to not create Order model when OrderItem or OrderItemAddition will be failed

* fix: update update_discount view to use Bill.add_discount